Revive SplitFile pass.
This (kind of) reverts commit eaa43850ff7a7ae14467ecb358bc392aef534d8d.

We still need to be able to specify different PPCG flags for
different functions that are originally in a single file, so we revive
the SplitFile pass.

The pass has been updated as follows over the dropped original:

 - Include printing of types.  Not sure how this was handled in the
   original pass.
 - Rename the overrided function "execute" into "walkProgram".
 - Only print functions that have a body and are not static.
 - Build a call graph, to include all called functions in the new
   files.

Change-Id: I4c152ce7fdbcf89f91a4990222c762d70b7ee86b